pyproject.toml
[tool.poetry]
name = "smartschool"
version = "0.1.0"
description = "Unofficial API interface to the smartschool system."
authors = ["Steven 'KaReL' Van Ingelgem <steven@vaningelgem.be>"]
readme = "README.md"
homepage = "https://github.com/svaningelgem/smartschool"
license = "GPL-3.0"

[tool.poetry.urls]
"Source Code" = "https://github.com/svaningelgem/smartschool"

[tool.poetry.dependencies]
python = "^3.10"
bs4 = "^0.0.1"
requests = "^2.31"
pyyaml = "^6.0"
pydantic = "^2.5"

[tool.poetry.group.dev.dependencies]
pytest-mock = "^3.12"
pytest-cov = "^4.1"
requests-mock = "^1.11"
ruff = "^0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
addopts = "--cov=smartschool --cov-append --cov-report=json --cov-fail-under=80 --cov-branch"
testpaths = [
    "tests",
]
python_files = ['*_tests.py', '*_test.py']
pythonpath = [
    "src"
]

[tool.ruff]
line-length = 160
fix = true

select = [
    "E", # pycodestyle
    "W", # pycodestyle
    "F", # Pyflakes
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "D", # flake8-docstrings
    "I", # isort
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "TID", # flake8-tidy-imports
    "UP", # pyupgrade
    "PT", # flake8-pytest-style
    "RUF", # Ruff-specific rules
    "PTH", # flake8-use-pathlib
    "FA", # flake8-future-annotations
    "PIE", # flake8-pie
    "TD", # flake8-todos
    "TRY", # tryceratops
    "FBT001", # flake8-boolean-trap
]

ignore = [
    # Line length regulated by black
    "E501",
    # pydocstyle: http://www.pydocstyle.org/en/stable/error_codes.html
    # numpy convention with a few additional lints
    "D107",
    "D203",
    "D212",
    "D401",
    "D402",
    "D415",
    "D416",
    # flake8-pytest-style:
    "PT011", # pytest.raises({exception}) is too broad, set the match parameter or use a more specific exception
    # flake8-simplify
    "SIM102", # Use a single `if` statement instead of nested `if` statements
    "SIM108", # Use ternary operator
    # ruff
    "RUF005", # unpack-instead-of-concatenating-to-collection-literal
    # pycodestyle
    # TODO: Remove errors below to further improve docstring linting
    # Ordered from most common to least common errors.
    "D105", # Missing docstring in magic method
    "D100", # Missing docstring in public module
    "D104", # Missing docstring in public package
    # flake8-todos
    "TD002", # Missing author in TODO
    "TD003", # Missing issue link on the line following this TODO
    # tryceratops
    "TRY003", # Avoid specifying long messages outside the exception class
    # Lints below are turned off because of conflicts with the ruff formatter
    "D206",
    "W191",

    "TID252",  # Relative imports are banned
    "D101",  # Missing docstring in public class
    "D102",  # Missing docstring in public method
    "D103",  # Missing docstring in public function
]

[tool.ruff.pycodestyle]
max-doc-length = 160

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.flake8-type-checking]
strict = true

[tool.ruff.per-file-ignores]
"tests/**/*.py" = ["D100", "D103", "B018", "FBT001"]